Create ForwardPass class, break Backbufffer's Texture inheritance
ForwardPass class now contains draw code which resided in Renderer. This
was done to prevent code overflow while implementing rendering algorithms
in the project.

Introducing ForwardPass forced a requirement to render everything to
separate texture and copy results to Backbuffer. This allowed to break
Backbuffer's dependency from Texture class, which simplified a lot of
code.