Implement View Frustum Culling
To lighten the CPU, View Frustum Culling was implemented. It's pass
is performed once before waiting for fences from previous frame to
speed up the application and don't overdo the tests for each pass.

NOTE: Rotation in objects might not work - AABBs are only rotated
together with the object using same transform, NOT recalculated as
they should be.